fixed 继续点单bug修复

+++ fuintBackend/fuint-application/src/main/java/com/fuint/common/dto/HangUpDto.java
@@ -26,7 +26,7 @@ public class HangUpDto {
     private MtUser memberInfo;
 
     @ApiModelProperty("件数")
-    private Integer num;
+    private Double num;
 
     @ApiModelProperty("金额")
     private BigDecimal amount;
